854S01AKILF Hidden symbol pins!
Hi, not too sure you will be happy with my workaround for this but whoever made the schematic symbol has hidden the power and gnd pins. GND is fine but the "VDD" rail will not short to the 3V3 provided. I have altered settings within Concept HDL to let me use a synonym (which does work but not on the logical net set by the symbol) to short the 3.3V and VDD but it doesn't work. I have used a ferrite bead to join the voltages together and some bypass caps.